last version

+++ PlataTaxe/src/java/models/AutentificareJpaController.java
@@ -75,7 +75,7 @@ public class AutentificareJpaController implements Serializable {
             }
             String msg = ex.getLocalizedMessage();
             if (msg == null || msg.length() == 0) {
-                Integer id = autentificare.getId();
+                Long id = autentificare.getId();
                 if (findAutentificare(id) == null) {
                     throw new NonexistentEntityException("The autentificare with id " + id + " no longer exists.");
                 }
@@ -140,7 +140,7 @@ public class AutentificareJpaController implements Serializable {
         }
     }
 
-    public Autentificare findAutentificare(Integer id) {
+    public Autentificare findAutentificare(Long id) {
         EntityManager em = getEntityManager();
         try {
             return em.find(Autentificare.class, id);
